sloane: Fix btwrite timer timeout
Set PROC_BTWRITE_TIMER_TIMEOUT_MS to 0
to avoid upio start and stop timer when LMP
is enabled and BT_WAKE is asserted and
de-asserted. This time is useless and would
set btwrite to 0 by mistake even when host
is sending data to UART, and this would
de-assert BT_WAKE by mistake.

Bug: 23248798
Change-Id: Ie55c346785679e51a771d5e46270c0858e6082ee

Signed-off-by: jmpfbmx <jmpf.bmx@gmail.com>